bit of a refactor
