+++ products/signals/backend/management/commands/signal_pipeline_status.py
@@ -4,8 +4,6 @@ import asyncio
 
 from django.core.management.base import BaseCommand, CommandError
 
-from posthog.schema import EmbeddingModelName
-
 from posthog.hogql import ast
 from posthog.hogql.query import execute_hogql_query
 
@@ -14,8 +12,7 @@ from posthog.models import Team
 from products.signals.backend.models import SignalReport
 from products.signals.backend.temporal.grouping import TeamSignalGroupingWorkflow
 from products.signals.backend.temporal.summary import SignalReportSummaryWorkflow
-
-EMBEDDING_MODEL = EmbeddingModelName.TEXT_EMBEDDING_3_SMALL_1536
+from products.signals.backend.utils import EMBEDDING_MODEL
 
 
 class Command(BaseCommand):
